+++ examples/named_cluster/main.tf
@@ -74,6 +74,7 @@ module "aks_cluster_name" {
   }
   net_profile_pod_cidr              = "10.1.0.0/16"
   private_cluster_enabled           = true
+  rbac_aad                          = true
   rbac_aad_managed                  = true
   role_based_access_control_enabled = true
 }