+++ examples/startup/main.tf
@@ -78,6 +78,7 @@ module "aks" {
   network_policy                    = "azure"
   os_disk_size_gb                   = 60
   private_cluster_enabled           = true
+  rbac_aad                          = true
   rbac_aad_managed                  = true
   role_based_access_control_enabled = true
   sku_tier                          = "Paid"